correcion en administrador e itemProducto donde se actualiza en el estado

+++ src/components/pages/producto/ItemProducto.jsx
@@ -1,25 +1,42 @@
 import { Button } from "react-bootstrap";
-import { eliminarProductoAPI } from "../../../helpers/queries";
+import { eliminarProductoAPI, obtenerProductosAPI } from "../../../helpers/queries";
 import Swal from "sweetalert2";
 
-const ItemProducto = ({producto}) => {
+const ItemProducto = ({producto,setProductos}) => {
 
   const eliminarProducto = async (id) => {
-    const respuesta = await eliminarProductoAPI(id);
-    if(respuesta.status === 200){
-      Swal.fire({
-        title: "Producto eliminado",
-        text: `El producto "${producto.nombreProducto}" fue eliminado correctamente`,
-        icon: "success"
-      });
-    }else{
-      Swal.fire({
-        title: "Ocurrio un error",
-        text: `El producto "${producto.nombreProducto}" no pudo ser eliminado. Intenta esta operación en unos minutos.`,
-        icon: "error"
-      });
-    }
+    // const respuesta = await eliminarProductoAPI(id);
+    Swal.fire({
+      title: "Estas seguro?",
+      text: "No puedes revertir esto!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, borrar"
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const respuesta = await eliminarProductoAPI(id);
+        Swal.fire({
+          title: "Borrado!",
+          text: "Su producto ha sido eliminado.",
+          icon: "success"
+        });
+        const repuestaProductos = await obtenerProductosAPI();
+          const productosRestantes = await repuestaProductos;
+          setProductos(productosRestantes);
+        }else{
+          Swal.fire({
+            title: "Ocurrio un error",
+            text: `El producto "${producto.nombreProducto}" no pudo ser eliminado. Intenta esta operación en unos minutos.`,
+            icon: "error"
+          });
+        }
+    });
+  
   }
+
+
 
   return (
     <tr>
